Release/v0.6.0

+++ portal/apps/vcs.go
@@ -13,9 +13,10 @@ import (
 	"cloudiac/portal/services/vcsrv"
 	"cloudiac/utils"
 	"fmt"
-	"github.com/gin-gonic/gin"
 	"net/http"
 	"strings"
+
+	"github.com/gin-gonic/gin"
 )
 
 func CreateVcs(c *ctx.ServiceContext, form *forms.CreateVcsForm) (interface{}, e.Error) {
@@ -84,6 +85,15 @@ func DeleteVcs(c *ctx.ServiceContext, form *forms.DeleteVcsForm) (result interfa
 	if err != nil {
 		return nil, err
 	}
+	// 根据vcsId查询是否相关云模版已经被全部清除
+	exist, err := services.QueryTplByVcsId(c.DB(), form.Id)
+	if err != nil {
+		return nil, err
+	}
+	if exist {
+		return nil, e.New(e.VcsDeleteError, fmt.Errorf("Vcs cannot be deleted. Please delete the dependent cloud template first"))
+	}
+
 	if err := services.DeleteVcs(c.DB(), form.Id); err != nil {
 		return nil, err
 	}
@@ -286,4 +296,30 @@ func VcsVariableSearch(c *ctx.ServiceContext, form *forms.TemplateVariableSearch
 	}
 
 	return tvl, nil
+}
+
+func SearchVcsFile(c *ctx.ServiceContext, form *forms.SearchVcsFileForm) (interface{}, e.Error) {
+	vcs, err := checkOrgVcsAuth(c, form.Id)
+	if err != nil {
+		return nil, err
+	}
+	vcsService, er := vcsrv.GetVcsInstance(vcs)
+	if er != nil {
+		return nil, e.New(e.VcsError, er)
+	}
+	repo, er := vcsService.GetRepo(form.RepoId)
+	if er != nil {
+		return nil, e.New(e.VcsError, er)
+	}
+	b, er := repo.ReadFileContent(form.Branch, form.FileName)
+	if er != nil {
+		if strings.Contains(er.Error(), "not found") {
+			b = make([]byte, 0)
+		} else {
+			return nil, e.New(e.VcsError, er)
+		}
+	}
+
+	res := gin.H{"content": string(b)}
+	return res, nil
 }
